backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic-settings
torch
torchvision
pillow
numpy
scikit-learn
sqlalchemy
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
google-genai
supabase
psycopg2-binary
pytest
httpx